fixing bug where compontents folder was affecting changes to files
